Added FileLister class and VideoData class
VideoData class was renamed from VideoInfo because it had the same name as a class in the Dropbox API.

Video buttons and descriptions are generated in the video gallery.

View Video screen uses the VideoData object to play video.

Removed AppStartup activity because it was redundant.

Added a wait period of 10 seconds for the Dropbox API to get the Videos. (For Home and VideoGallery Activities).